# openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon Platform API
  description: |
    REST API веб-платформы для проведения хакатонов.
    Курсовой проект: клиент-серверное приложение (Go + React + PostgreSQL).

    **Домен:** хакатон - соревнование команд за ограниченное время; результат - прототип + демо + питч.
    Платформа покрывает: публикацию события, треки/кейсы, регистрацию, команды с ролями, сдачу артефактов до дедлайна.

    Полная спецификация страниц и use-case: `docs_coursework/TECHNICAL_IMPLEMENTATION.md`
  version: 0.1.0
  contact:
    name: PiRKSPCoursework
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.example.com/api/v1
    description: Production (placeholder)
...
paths:
...
  /hackathons/{hackathonId}/register:
    post:
      tags: [Participation]
      summary: Записаться на хакатон (индивидуальная регистрация)
      description: |
        Доступно в статусе `registration` и до `registration_closes_at`.
        После этого участник может создать или вступить в команду.
      operationId: registerForHackathon
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/HackathonId"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HackathonRegistration"
        "400":
          description: Регистрация закрыта или уже зарегистрирован
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Только роль participant (organizer не регистрируется как участник)
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Participation]
      summary: Отменить регистрацию на хакатон
      description: Нельзя, если пользователь уже в команде с сабмитом (упрощённо - нельзя в команде).
      operationId: unregisterFromHackathon
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/HackathonId"
      responses:
        "204":
          description: Регистрация отменена
        "400":
          description: Нельзя отменить (уже в команде)
        "401":
          $ref: "#/components/responses/Unauthorized"
  /hackathons/{hackathonId}/teams:
    get:
      tags: [Teams]
      summary: Список команд хакатона
      description: Публичный список названий команд (без PII участников) или с составом - на усмотрение MVP (здесь - с составом для прозрачности).
      operationId: listTeams
      parameters:
        - $ref: "#/components/parameters/HackathonId"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Teams]
      summary: Создать команду
      description: |
        Создатель становится капитаном и участником с ролью `team_lead`.
        Требуется предварительная регистрация на хакатон.
      operationId: createTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/HackathonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Уже в команде / лимит команд / регистрация закрыта
        "401":
          $ref: "#/components/responses/Unauthorized"
  /teams/{teamId}:
    get:
      tags: [Teams]
      summary: Карточка команды
      operationId: getTeam
      parameters:
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Teams]
      summary: Обновить команду (капитан)
      description: Смена названия, трека, кейса. Только капитан.
      operationId: updateTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TeamId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "403":
          description: Не капитан
        "401":
          $ref: "#/components/responses/Unauthorized"
  /teams/{teamId}/join:
    post:
      tags: [Teams]
      summary: Вступить в команду
      description: |
        Проверяется `max_team_size` хакатона. Участник должен быть зарегистрирован на хакатон.
        Роль в команде можно передать сразу или выставить позже через PATCH member.
      operationId: joinTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TeamId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinTeamRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Команда полна / уже в другой команде
        "401":
          $ref: "#/components/responses/Unauthorized"
  /teams/{teamId}/leave:
    post:
      tags: [Teams]
      summary: Покинуть команду
      description: Капитан не может выйти, не передав капитанство (MVP - запрет выхода капитана).
      operationId: leaveTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TeamId"
      responses:
        "204":
          description: Вышел из команды
        "400":
          description: Капитан не может выйти без расформирования
  /teams/{teamId}/members/{userId}:
    patch:
      tags: [Teams]
      summary: Обновить роль участника в команде
      description: Себя или капитан меняет любого.
      operationId: updateTeamMemberRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TeamId"
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [team_role]
              properties:
                team_role:
                  $ref: "#/components/schemas/TeamMemberRole"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "403":
          description: Нет прав
  /teams/{teamId}/submission:
    get:
      tags: [Submissions]
      summary: Сабмит команды
      operationId: getTeamSubmission
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Сабмит ещё не создан
    put:
      tags: [Submissions]
      summary: Создать или обновить сабмит
      description: |
        Доступно участникам команды до `submission_deadline_at` хакатона.
        Артефакты: репозиторий (обяз.), демо, питч, видео (опц.).
      operationId: upsertTeamSubmission
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TeamId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertSubmissionRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          description: Дедлайн прошёл / не выбран кейс
        "403":
          description: Не член команды
...
  /organizer/hackathons/{hackathonId}/submissions:
    get:
      tags: [Organizer, Submissions]
      summary: Все сабмиты хакатона (для организатора / «жюри»)
      description: Просмотр артефактов без выставления оценок в MVP.
      operationId: listHackathonSubmissions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/HackathonId"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SubmissionWithTeam"
  /organizer/hackathons/{hackathonId}/registrations:
    get:
      tags: [Organizer]
      summary: Список зарегистрированных участников
      operationId: listHackathonRegistrations
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/HackathonId"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/HackathonRegistrationWithUser"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    HackathonId:
      name: hackathonId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TeamId:
      name: teamId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
    PlatformRole:
      type: string
      enum: [participant, organizer]
      description: Роль аккаунта на платформе
    TeamMemberRole:
      type: string
      enum:
        - team_lead
        - developer
        - designer
        - data_scientist
        - devops_qa
        - other
      description: Роль участника внутри команды (не нормируется строго, для состава питча)
...
    User:
      type: object
      required: [id, email, full_name, platform_role, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        platform_role:
          $ref: "#/components/schemas/PlatformRole"
        created_at:
          type: string
          format: date-time
...
    HackathonRegistration:
      type: object
      required: [id, hackathon_id, user_id, registered_at]
      properties:
        id:
          type: string
          format: uuid
        hackathon_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        registered_at:
          type: string
          format: date-time
    HackathonRegistrationWithUser:
      allOf:
        - $ref: "#/components/schemas/HackathonRegistration"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
...
    Team:
      type: object
      required: [id, hackathon_id, name, captain_id, members]
      properties:
        id:
          type: string
          format: uuid
        hackathon_id:
          type: string
          format: uuid
        name:
          type: string
        captain_id:
          type: string
          format: uuid
        track_id:
          type: string
          format: uuid
          nullable: true
        case_id:
          type: string
          format: uuid
          nullable: true
        members:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
    TeamMember:
      type: object
      required: [user_id, full_name, team_role]
      properties:
        user_id:
          type: string
          format: uuid
        full_name:
          type: string
        team_role:
          $ref: "#/components/schemas/TeamMemberRole"
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        track_id:
          type: string
          format: uuid
        case_id:
          type: string
          format: uuid
        team_role:
          $ref: "#/components/schemas/TeamMemberRole"
          default: team_lead
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
        track_id:
          type: string
          format: uuid
        case_id:
          type: string
          format: uuid
    JoinTeamRequest:
      type: object
      properties:
        team_role:
          $ref: "#/components/schemas/TeamMemberRole"
          default: developer
    Submission:
      type: object
      required: [id, team_id, hackathon_id, repo_url, updated_at]
      properties:
        id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
        hackathon_id:
          type: string
          format: uuid
        title:
          type: string
        summary:
          type: string
          description: Краткое описание решения
        repo_url:
          type: string
          format: uri
        demo_url:
          type: string
          format: uri
        pitch_url:
          type: string
          format: uri
          description: Презентация (PDF / Google Slides)
        video_url:
          type: string
          format: uri
        submitted_at:
          type: string
          format: date-time
          description: Первичная сдача (null до первого save)
        updated_at:
          type: string
          format: date-time
    SubmissionWithTeam:
      allOf:
        - $ref: "#/components/schemas/Submission"
        - type: object
          properties:
            team_name:
              type: string
            case_title:
              type: string
            track_title:
              type: string
    UpsertSubmissionRequest:
      type: object
      required: [repo_url]
      properties:
        title:
          type: string
        summary:
          type: string
        repo_url:
          type: string
          format: uri
        demo_url:
          type: string
          format: uri
        pitch_url:
          type: string
          format: uri
        video_url:
          type: string
          format: uri